feat: 添加使用scout实现搜索结果展示及分页

+++ app/Http/Controllers/PostController.php
@@ -2,23 +2,25 @@
 
 namespace App\Http\Controllers;
 
-use Illuminate\Http\Request;
+use App\Comment;
 use \App\Post;
+use \App\Zan;
+use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 class PostController extends Controller
 {
     // 列表页面
-    public function index(\Psr\Log\LoggerInterface $log)
+    public function index()
     {
-        \Log::info("post_index", ['data' => 'this is post index']);
-
-        $posts = Post::orderBy('created_at', 'desc')->paginate(6);
+        $posts = Post::orderBy('created_at', 'desc')->withCount(["comments", "zans"])->paginate(6);
         return view("post/index", compact('posts'));
     }
 
     // 详情页面
     public function show(Post $post)
     {
+        $post->load('comments');
         return view("post/show", compact('post'));
     }
 
@@ -61,6 +63,8 @@ class PostController extends Controller
             'content' => 'required|string|min:10'
         ]);
 
+        $this->authorize('update', $post);
+
         // 逻辑
         $post->title = request('title');
         $post->content = request('content');
@@ -73,7 +77,7 @@ class PostController extends Controller
     // 删除逻辑
     public function delete(Post $post)
     {
-        // TODO: 用户权限的认证
+        $this->authorize('delete', $post);
         $post->delete();
         return redirect("/posts");
     }
@@ -83,5 +87,59 @@ class PostController extends Controller
     {
         $path = $request->file('wangEditorH5File')->storePublicly(md5(time()));
         return asset('storage/' . $path);
+    }
+
+    // 提交评论
+    public function comment(Post $post)
+    {
+        // 验证
+        $this->validate(request(), [
+            'content' => 'required|min:3',
+        ]);
+
+        // 逻辑
+        $comment = new Comment();
+        $comment->user_id = Auth::id();
+        $comment->content = request('content');
+        $post->comments()->save($comment);
+
+        // 渲染
+        return back();
+    }
+
+    // 赞
+    public function zan(Post $post)
+    {
+        $param = [
+            'user_id' => Auth::id(),
+            'post_id' => $post->id,
+        ];
+
+        Zan::firstOrCreate($param);
+
+        return back();
+    }
+
+    // 取消赞
+    public function unzan(Post $post)
+    {
+        $post->zan(Auth::id())->delete();
+        return back();
+    }
+
+    // 搜索结果页
+    public function search()
+    {
+        // 验证
+        $this->validate(request(), [
+            'query' => 'required'
+        ]);
+//
+        // 逻辑
+        $query = request('query');
+        $posts = Post::search($query)->paginate(2);
+//
+        // 渲染
+        return view("post/search", compact('posts', 'query'));
     }
 }
